Improve type name mapping to handle unions
Rather than naively stringifying types, intelligently map from types
to names, including a normalization pass for unions.

This does lead to potential undefined behavior if a user passes a
Union and UnionType which are synonyms, but that's sufficiently weird
and incorrect usage to allow as undefined.

This replaces any notion of an external type-to-name mapping.