Use callback and convert return types
For both of these, evaluate type annotations using
`typing.get_type_hints()`, which requires that we find and use the
module global namespace for whatever module defined the calls.

Having done so, it is possible to check these return types. Restrict
the use of `convert` logic to non-`click` definitions, since `click`
frequently defines `Any` and we'd rather use the `click` classes. But
the check precedes use of the `click` classes so that it will work on
custom types which inherit from `click` *if* they override the default
`convert` method.