Add support for click version 8.3+
The `default` attribute of options and optional-arguments is now `UNSET`,
an internal sentinel defined for use within `click`.

Rather than checking if `default is not None` to see if there is a
"set default" on an argument, we now check
`default not in [UNSET, None]`.

A new test case covers an explicit `default=None`, which guarantees that
in future versions, we won't start assuming that only `default=UNSET`
indicates nullability.